+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/Shared/PrometheusExporter.cs
@@ -27,6 +27,7 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
         this.ScrapeResponseCacheDurationMilliseconds = options.ScrapeResponseCacheDurationMilliseconds;
         this.TargetInfoEnabled = options.TargetInfoEnabled;
         this.DisableTotalNameSuffixForCounters = options.DisableTotalNameSuffixForCounters;
+        this.ResourceConstantLabels = options.ResourceConstantLabels;
         this.MaxScrapeResponseSizeBytes = options.MaxScrapeResponseSizeBytes;
 
         this.CollectionManager = new PrometheusCollectionManager(this);
@@ -52,6 +53,8 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
     internal int ScrapeResponseCacheDurationMilliseconds { get; }
 
     internal bool DisableTotalNameSuffixForCounters { get; }
+
+    internal Func<string, bool>? ResourceConstantLabels { get; }
 
     internal int MaxScrapeResponseSizeBytes { get; }