+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -35,6 +35,11 @@ Notes](../../RELEASENOTES.md).
   when negotiated via the `Accept` header.
   ([#7440](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7440))
 
+* Add `PrometheusAspNetCoreOptions.ResourceConstantLabels` property to select
+  resource attributes to add to each metric as constant labels. Defaults to
+  `null` (no resource attributes are added as metric labels).
+  ([#7471](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7471))
+
 * Add `PrometheusAspNetCoreOptions.MaxScrapeResponseSizeBytes` to configure
   the maximum size of a scrape response. The default is now ~166 MiB.
   ([#7487](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7487))